+++ src/feature/my/component/index.js
@@ -2,5 +2,10 @@ export { default as AccountTab } from './AccountTab/AccountTab';
 export { default as ActivityTab } from './ActivityTab/ActivityTab';
 export { default as CircleProfile } from './CircleProfile/CircleProfile';
 export { default as MyInfo } from './MyInfo/MyInfo';
+export { default as MyPostList } from './MyPostList/MyPostList';
+export { default as MyPostListErrorFallback } from './MyPostList/MyPostListErrorFallback';
+export { default as PointLog } from './PointLog/PointLog';
+export { default as PointLogs } from './PointLogs/PointLogs';
+export { default as PointLogsErrorFallback } from './PointLogs/PointLogsErrorFallback';
 export { default as PolicyTab } from './PolicyTab/PolicyTab';
 export { default as TopOverlay } from './TopOverlay/TopOverlay';
