+++ src/page/user/index.js
@@ -1,0 +1,6 @@
+export { default as ActivityPage } from './ActivityPage/ActivityPage';
+export { default as ChangePwPage } from './ChangePwPage/ChangePwPage';
+export { default as DeleteAccountPage } from './DeleteAccountPage/DeleteAccountPage';
+export { default as EditProfilePage } from './EditProfilePage/EditProfilePage';
+export { default as MyPage } from './MyPage/MyPage';
+export { default as PointLogListPage } from './PointLogListPage/PointLogListPage';
