+++ src/pages/MyPage/index.js
@@ -1,9 +1,2 @@
-export { default as MyPage } from './MyPage';
-
-export { default as PointLogsPage } from './pages/PointLogsPage/PointLogsPage';
-export { default as ActivityPage } from './pages/ActivityPage/ActivityPage';
-export { default as EditInfoPage } from './pages/EditInfoPage/EditInfoPage';
-export { default as ChangePasswordPage } from './pages/ChangePasswordPage/ChangePasswordPage';
-export { default as DeleteAccountPage } from './pages/DeleteAccountPage/DeleteAccountPage';
-export { default as PrivacyPolicyPage } from './pages/PolicyPage/PrivacyPolicyPage';
-export { default as ServicePolicyPage } from './pages/PolicyPage/ServicePolicyPage';
+export { default as PrivacyPolicyPage } from './PolicyPage/PrivacyPolicyPage';
+export { default as ServicePolicyPage } from './PolicyPage/ServicePolicyPage';
