--- src/pages/MyPage/pages/ActivityPage/index.js
@@ -1,4 +1,0 @@
-export { default as ActivityPage } from './ActivityPage';
-
-export { default as Posts } from './components/Posts/Posts';
-export { default as PostsErrorFallback } from './components/Posts/PostsErrorFallback';
