--- src/pages/MyPage/pages/PointLogsPage/index.js
@@ -1,5 +1,0 @@
-export { default as PointLogsPage } from './PointLogsPage';
-
-export { default as PointLog } from './components/PointLog/PointLog';
-export { default as PointLogs } from './components/PointLogs/PointLogs';
-export { default as PointLogsErrorFallback } from './components/PointLogs/PointLogsErrorFallback';
